app/requirements.txt
-i https://pypi.org/simple
asgiref==3.3.0; python_version >= '3.5'
django==3.1.3
djangorestframework-simplejwt==4.5.0
djangorestframework==3.12.2
pillow==8.0.1
psycopg2-binary==2.8.6
pyjwt==1.7.1
pytz==2020.4
sqlparse==0.4.1; python_version >= '3.5'
